[feature] bug fix & combine namuApi
